+++ src/test/java/integration/BaseIntegrationTest.java
@@ -11,17 +11,18 @@ import java.util.Locale;
 
 import static com.automation.remarks.video.enums.RecordingMode.ANNOTATED;
 import static com.codeborne.selenide.Browsers.FIREFOX;
+import static com.codeborne.selenide.Browsers.SAFARI;
 import static com.codeborne.selenide.impl.FileHelper.ensureFolderExists;
 import static java.lang.Boolean.parseBoolean;
 import static org.openqa.selenium.net.PortProber.findFreePort;
 
 @ExtendWith({LogTestNameExtension.class})
 public abstract class BaseIntegrationTest {
-  private static final boolean SSL = true;
   protected static LocalHttpServer server;
   private static String protocol;
   private static int port;
   protected static final String browser = System.getProperty("selenide.browser", FIREFOX);
+  private static final boolean SSL = !SAFARI.equalsIgnoreCase(browser);
   static final boolean headless = parseBoolean(System.getProperty("selenide.headless", "false"));
 
   @BeforeAll