+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 5.14.0
+* #1236 add support for Safari browser  --  see PR #1237
 * #1220 create a unique downloads folder for every browser instance  --  see PR #1221
 * #1194 added method `$$.shouldHave(itemWithText("any text"))`  --  thanks to Luis Serna for PR #1194
 * #1166 added method `SelenideDriver.screenshot(fileName)`  --  see PR #1227